rename molgraph to graph
