+++ mol_gnn/data/molecule.py
@@ -6,9 +6,9 @@ from rdkit import Chem
 from sklearn.preprocessing import StandardScaler
 from torch.utils.data import Dataset
 
+from mol_gnn.types import Mol
 from mol_gnn.data.mixins import _DatapointMixin, _MolGraphDatasetMixin, Datum
-from mol_gnn.featurizers.molecule import MoleculeFeaturizer
-from mol_gnn.featurizers.molgraph import MoleculeMolGraphFeaturizer, BaseMoleculeMolGraphFeaturizer
+from mol_gnn.featurizers import VectorFeaturizer, GraphFeaturizer, SimpleMoleculeMolGraphFeaturizer
 from mol_gnn.utils.chem import make_mol
 
 
@@ -44,7 +44,7 @@ class MoleculeDatapoint(_DatapointMixin, _MoleculeDatapointMixin):
     ``d_vd`` is the number of additional features that will be concatenated to atom-level features
     *after* message passing"""
 
-    def __post_init__(self, mfs: list[MoleculeFeaturizer] | None):
+    def __post_init__(self, mfs: list[VectorFeaturizer[Mol]] | None):
         if self.mol is None:
             raise ValueError("Input molecule was `None`!")
 
@@ -59,7 +59,7 @@ class MoleculeDatapoint(_DatapointMixin, _MoleculeDatapointMixin):
 
         super().__post_init__(mfs)
 
-    def calc_features(self, mfs: list[MoleculeFeaturizer]) -> np.ndarray:
+    def calc_features(self, mfs: list[VectorFeaturizer[Mol]]) -> np.ndarray:
         if self.mol.GetNumHeavyAtoms() == 0:
             return np.zeros(sum(len(mf) for mf in mfs))
 
@@ -74,12 +74,12 @@ class MoleculeDataset(_MolGraphDatasetMixin, Dataset):
     ----------
     data : Iterable[MoleculeDatapoint]
         the data from which to create a dataset
-    featurizer : MoleculeFeaturizer
+    featurizer : Featurizer[Mol]
         the featurizer with which to generate MolGraphs of the molecules
     """
 
     data: list[MoleculeDatapoint]
-    featurizer: MoleculeMolGraphFeaturizer = field(default_factory=BaseMoleculeMolGraphFeaturizer())
+    featurizer: GraphFeaturizer[Mol] = field(default_factory=SimpleMoleculeMolGraphFeaturizer)
 
     def __getitem__(self, idx: int) -> Datum:
         d = self.data[idx]
